+++ src/RecoveryServices/RecoveryServices.Backup.Models/CmdletParamEnums.cs
@@ -58,7 +58,10 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models
     public enum RestoreVMBackupItemParams
     {
         TargetResourceGroupName,
-        OsaOption
+        OsaOption,
+        RestoreDiskList,
+        RestoreOnlyOSDisk,
+        RestoreAsUnmanagedDisks
     }
 
     public enum RestoreFSBackupItemParams
@@ -118,7 +121,11 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models
         BackupType,
         EnableCompression,
         DeleteState,
-        FriendlyName
+        FriendlyName,
+        InclusionDisksList,
+        ExclusionDisksList,
+        ResetExclusionSettings,
+        ExcludeAllDataDisks
     }
 
     public enum ProtectionCheckParams