+++ src/RecoveryServices/RecoveryServices.Backup/ParamHelpMsgs.cs
@@ -95,6 +95,10 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
             public const string EnableCompression = "A switch which will specify that the requested on-demand SQL backup should be compressed.";
             public const string ParentID = "Specified the ARM ID of an Instance or AG.";
             public const string FriendlyName = "FriendlyName of the backed up item";
+            public const string inclusionDiskList = "List of Disk LUNs to include in backup";
+            public const string exclusionDiskList = "List of Disk LUNs to exclude in backup";
+            public const string resetExclusionSettings = "Specifies to reset disk exclusion setting associated with the item";
+            public const string excludeAllDataDisks = "Option to specify to backup OS disks only";
         }
 
         internal static class ProtectableItem
@@ -138,6 +142,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets
         {
             public const string TargetResourceGroupName = "The resource group to which the managed disks are restored. This parameter is mandatory for backup of VM with managed disks";
             public const string OsaOption = "Use this switch if the disks from the recovery point are to be restored to their original storage accounts";
+            public const string RestoreOnlyOSDisk = "Use this switch to restore only OS disks of a backed up VM";
+            public const string RestoreDiskList = "Specify which disks to recover of the backed up VM";
+            public const string RestoreAsUnmanagedDisks = "Use this switch to specify to restore as unmanaged disks";
         }
 
         internal static class RestoreFS